+++ external/ann_1.1.1/src/kd_pr_search.h
@@ -38,10 +38,10 @@
 //		Appx_k_Near_Neigh().
 //----------------------------------------------------------------------
 
-extern double			ANNprEps;		// the error bound
+extern float			ANNprEps;		// the error bound
 extern int				ANNprDim;		// dimension of space
 extern ANNpoint			ANNprQ;			// query point
-extern double			ANNprMaxErr;	// max tolerable squared error
+extern float			ANNprMaxErr;	// max tolerable squared error
 extern ANNpointArray	ANNprPts;		// the points
 extern ANNpr_queue		*ANNprBoxPQ;	// priority queue for boxes
 extern ANNmin_k			*ANNprPointMK;	// set of k closest points
